rename dir to original name
