+++ Sivatagi vizhalozatok uzemeltetese a gyakorlatban 2/src/Controll/Controller.java
