+++ Sivatagi vizhalozatok uzemeltetese a gyakorlatban 2/src/Drawing/Drawable.java
