+++ Sivatagi vizhalozatok uzemeltetese a gyakorlatban 2/src/Drawing/SaboteurDraw.java
