+++ Sivatagi vizhalozatok uzemeltetese a gyakorlatban 2/src/Enums/Fluid.java
