+++ Sivatagi vizhalozatok uzemeltetese a gyakorlatban 2/src/Fields/ActiveFields/ActiveFields.java
