+++ Sivatagi vizhalozatok uzemeltetese a gyakorlatban 2/src/Fields/Field.java
