+++ Sivatagi vizhalozatok uzemeltetese a gyakorlatban 2/src/Interfaces/Steppable.java
