+++ Sivatagi vizhalozatok uzemeltetese a gyakorlatban 2/src/Players/Mechanic.java
