+++ Sivatagi vizhalozatok uzemeltetese a gyakorlatban 2/src/Players/Player.java
