+++ Sivatagi vizhalozatok uzemeltetese a gyakorlatban 2/src/palya.txt
