--- Sivatagi_vizhalozatok_uzemeltetese_a_gyakorlatban_2/src/StringResource/StringResourceController.java
@@ -1,13 +1,0 @@
-package StringResource;
-
-public final class StringResourceController {
-    private StringResourceController(){}
-    public static final String FALSE = "false";
-    public static final String FILE_NOT_FOUND = "File not found";
-    public static final String WRONG_PLAYER = "Nem te vagy a soron következő játékos!";
-    public static final String WRONG_ACTION = "Sikertelen művelet";
-    public static final String GOOD_ACTION = "Sikeres művelet";
-    public static final String WATER = "water";
-    public static final String RANDOM_ON = "A véletlen események be lettek kapcsolva.";
-
-}
